+++ lua/AI/LobbyOptions/LobbyOptions.lua
@@ -57,4 +57,15 @@ AIOpts = {
 			'1.0', '1.25', '1.5', '1.75', '2.0', '2.25', '2.5', '2.75', '3.0', '3.25', '3.5', '3.75', '4.0', '4.5', '5.0', '5.5', '6.0', '6.5', '7.0', '7.5', '8.0',
         },
     },
+	{
+        default = 1,
+        label = "Campaign AI Construction Delay:",
+        help = "Set the delay for AI Engineering units when to build another structure once it\'s current structure has been finished, in game ticks. 10 game ticks translates to 1 second, default is 10",
+		key = 'EngineerBuildDelay',
+		value_text = "%s",
+        value_help = "Build delay of %s",
+        values = {
+			'10', '15', '20', '25', '30', '35', '40', '45', '50', '55', '60', '65', '70', '75', '80', '85', '90', '95', '100',
+        },
+    },
 }