+++ mod_info.lua
@@ -1,6 +1,6 @@
 name = "Coop AI Adjustments"
-uid = "coop_ai_mod_v10"
-version = 10
+uid = "coop_ai_mod_v11"
+version = 11
 copyright = "MIT License"
 description = "Campaign AI fixes, adjustments, additions, and QoL changes, still WIP."
 author = "Dhomie"